MaxiSavings enhacements completed
Account.cs
Rates and percentages to be stored in a const variable.
Enhanced code readablity on calculating interest.
Enhancement for MaxiSavings: Changed to have an interest rate of 5%
assuming no withdrawals in the past. Did this by checking a past
withdrawal if it exists based off on its transaction date
CustomerTest
Included an Ignored 3 accounts test after completing enhancements on
maxisavings